pyproject.toml
[tool.poetry]
name = "stash"
version = "0.1.0"
description = "A tiny wishlist app for collecting product links and notes"
authors = ["Steven de Jong <steven@steven-dejong.nl>"]
readme = "README.md"
license = "MIT"
packages = [{ include = "stash_app", from = "." }]

[tool.poetry.dependencies]
python = "^3.13"
flask = "^2.3"
gunicorn = "^20.1"
flask-sqlalchemy = "^3.0"
flask-migrate = "^4.0"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0"
requests = "^2.32.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
